refactor(suite): extract message system sections into subcomponents
