+++ packages/suite/src/views/settings/SettingsDebug/MessageSystem/MessageSystemManager/MessageSystemManagerCondition.tsx
@@ -1,0 +1,35 @@
+import { IconName, InfoItem } from '@trezor/components';
+import { capitalizeFirstLetter } from '@trezor/utils';
+
+type MessageSystemManagerConditionProps<T extends Record<string, unknown>> = {
+    label: string;
+    iconName: IconName;
+    data?: T | readonly T[] | null;
+};
+
+export const MessageSystemManagerCondition = <T extends Record<string, unknown>>({
+    label,
+    iconName,
+    data,
+}: MessageSystemManagerConditionProps<T>) => {
+    // eslint-disable-next-line no-nested-ternary
+    const list: ReadonlyArray<T> = data == null ? [] : Array.isArray(data) ? data : [data];
+
+    return (
+        <InfoItem label={label} iconName={iconName} variant="default">
+            {list.length === 0
+                ? '-'
+                : list.flatMap((item, i) =>
+                      Object.entries(item).map(([key, value]) => (
+                          <InfoItem
+                              key={`${i}-${key}`}
+                              label={capitalizeFirstLetter(key)}
+                              direction="row"
+                          >
+                              {String(value)}
+                          </InfoItem>
+                      )),
+                  )}
+        </InfoItem>
+    );
+};
