+++ packages/suite/src/views/settings/SettingsDebug/SettingsDebug.tsx
@@ -18,8 +18,8 @@ import { FirmwareUpdate } from './FirmwareUpdate';
 import { ForgetAllDevicesButton } from './ForgetBluetoothDevices';
 import { GithubIssue } from './GithubIssue';
 import { InvityApi } from './InvityApi';
+import { MessageSystemConfigSourceSelect } from './MessageSystem/MessageSystemConfigSourceSelect';
 import { MessageSystemDebugInfo } from './MessageSystem/MessageSystemDebugInfo';
-import { MessageSystemConfigSourceSelect } from './MessageSystemConfigSourceSelect';
 import { Metadata } from './Metadata';
 import { OAuthApi } from './OAuthApi';
 import { PreField } from './PreField';
