+++ suite-common/message-system/src/__fixtures__/messageSystemReducer.ts
@@ -32,6 +32,8 @@ const initialState = {
     dismissedMessages: {},
 
     validExperiments: [],
+
+    manuallyAddedMessageIds: {},
 };
 
 export const fixtures = [
