+++ suite-common/message-system/src/messageSystemReducer.ts
@@ -22,7 +22,7 @@ const initialState: MessageSystemState = {
 
     configSource: 'remote',
 
-    inAppIds: {},
+    manuallyAddedMessageIds: {},
 };
 
 export const messageSystemInitialState = initialState;
@@ -60,13 +60,24 @@ export const prepareMessageSystemReducer = createReducerWithExtraDeps(
                 state.timestamp = timestamp;
                 state.currentSequence = config.sequence;
 
-                const prevInAppActions =
-                    state.config?.actions?.filter(a => !!state.inAppIds[a.message.id]) ?? [];
+                const validMessageIds = new Set(
+                    state.config?.actions?.map(a => a.message.id) ?? [],
+                );
+                state.manuallyAddedMessageIds = Object.fromEntries(
+                    Object.keys(state.manuallyAddedMessageIds ?? {})
+                        .filter(id => validMessageIds.has(id))
+                        .map(id => [id, true] as const),
+                );
+
+                const prevManuallyAddedMessageIds =
+                    state.config?.actions?.filter(
+                        action => !!state.manuallyAddedMessageIds[action.message.id],
+                    ) ?? [];
                 const incomingFileActions = config.actions ?? [];
 
                 state.config = {
                     ...config,
-                    actions: [...prevInAppActions, ...incomingFileActions],
+                    actions: [...prevManuallyAddedMessageIds, ...incomingFileActions],
                 };
             })
             .addCase(messageSystemActions.fetchError, state => {
@@ -89,7 +100,8 @@ export const prepareMessageSystemReducer = createReducerWithExtraDeps(
             .addCase(messageSystemActions.addMessage, (state, { payload }) => {
                 if (state.config) {
                     state.config.actions = [payload, ...state.config.actions];
-                    state.inAppIds[payload.message.id] = true;
+                    state.manuallyAddedMessageIds = state.manuallyAddedMessageIds || {};
+                    state.manuallyAddedMessageIds[payload.message.id] = true;
                 }
             })
             .addCase(messageSystemActions.removeMessage, (state, { payload }) => {
@@ -98,8 +110,8 @@ export const prepareMessageSystemReducer = createReducerWithExtraDeps(
                         action => action.message.id !== payload,
                     );
 
-                    if (state.inAppIds[payload]) {
-                        delete state.inAppIds[payload];
+                    if (state.manuallyAddedMessageIds[payload]) {
+                        delete state.manuallyAddedMessageIds[payload];
                     }
                 }
             })
