+++ suite-native/module-trading/src/selectors/__tests__/commonSelectors.test.ts
@@ -111,7 +111,7 @@ const getPreloadedState = ({
             dismissedMessages: {},
             validExperiments: [],
             configSource: 'local' as const,
-            inAppIds: {},
+            manuallyAddedMessageIds: {},
         },
     };
 };
